Well I fixed bug 87, albeit perhaps not in the most 'agreeable' way.
Heres the problem with fixing the bug in the IndexedTree class:
 - its used only internally
 - many classes (ie. SortedList) call get() hundreds of times, none of those are ever out-of-bounds
 - high performance is going to be a very important issue for SortedList, so we are going to avoid at all costs slowing that down

So I fixed it in CachingList instead. The rule is that IndexedTree will blow up violently if you violate its preconditions, so classes that use it should be careful to verify that they are not violating its preconditions.

Pretty much my idea is this:
  - create a user-friendly front end (ie. SortedList, CachingList) that does all the necessary due-process and range checks
  - create high performance backend (ie. IndexedTree, SparseList) that have strict requirements on parameters etc., but do no unnecessary checks or whatever.
  - in that high performance backend, use the assert() keyword to enable appropriate checking when necessary to verify that the code is infact correct


Issue number:  87
Obtained from:
Submitted by:
Reviewed by: